change files structure in app
