+++ app.js
@@ -4,8 +4,8 @@ const cors = require('cors');
 require("dotenv").config();
 
 const authRouter = require('./routes/api/auth');
+const usersRouter = require('./routes/api/users');
 const contactsRouter = require('./routes/api/contacts');
-const avatarsRouter = require('./routes/api/avatars');
 
 const app = express();
 
@@ -16,9 +16,9 @@ app.use(cors());
 app.use(express.json());
 app.use(express.static('public'));
 
-app.use('/api/users', authRouter);
+app.use('/api/auth', authRouter);
+app.use('/api/users', usersRouter);
 app.use('/api/contacts', contactsRouter);
-app.use('/avatars', avatarsRouter); // check express static work
 
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
