+++ controllers/users/index.js
@@ -1,0 +1,11 @@
+const logout = require('./logout');
+const current = require('./current');
+const updateSubscription = require('./updateSubscription');
+const updateAvatar = require('./updateAvatar');
+
+module.exports = {
+    logout,
+    current,
+    updateSubscription,
+    updateAvatar,
+};
